Added unit and integration tests for plugin
Holy moly, that was a lot of tests! Anyway, I've covered all the use cases I could think of. There might be more that escaped me, but for the most part, they are covered.

The tests mainly cover the Android file changing, as the iOS is covered by the Fastlane plugin `UpdateAppIdentifier`. If the iOS side fails, then it's down to the Fastlane plugin, not this one *successfully delegates responsibility*.